# spec/openapi-edukate-backend.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:5810
security:
- cookieAuth: []
...
paths:
...
  /api/v1/problem-sets:
    post:
      tags:
      - Problem Sets
      operationId: createProblemSet
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProblemSetRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemSetDto"
      security:
      - cookieAuth: []
  /api/v1/problem-sets/{shareCode}/visibility:
    post:
      tags:
      - Problem Sets
      operationId: changeVisibility
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      - name: isPublic
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemSetDto"
      security:
      - cookieAuth: []
  /api/v1/problem-sets/{shareCode}/role:
    post:
      tags:
      - Problem Sets
      operationId: changeUserRole
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      - name: username
        in: query
        required: true
        schema:
          type: string
          minLength: 1
      - name: requestedRole
        in: query
        required: true
        schema:
          type: string
          enum:
          - USER
          - MODERATOR
          - ADMIN
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
                enum:
                - USER
                - MODERATOR
                - ADMIN
      security:
      - cookieAuth: []
  /api/v1/problem-sets/{shareCode}/reply-invite:
    post:
      tags:
      - Problem Sets
      operationId: replyToInvite
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      - name: response
        in: query
        required: true
        schema:
          type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
      security:
      - cookieAuth: []
  /api/v1/problem-sets/{shareCode}/problems:
    post:
      tags:
      - Problem Sets
      operationId: changeProblems
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeProblemSetProblemsRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemSetDto"
      security:
      - cookieAuth: []
  /api/v1/problem-sets/{shareCode}/leave:
    post:
      tags:
      - Problem Sets
      operationId: leaveProblemSet
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
      security:
      - cookieAuth: []
  /api/v1/problem-sets/{shareCode}/invite:
    post:
      tags:
      - Problem Sets
      operationId: inviteToProblemSet
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      - name: inviteeName
        in: query
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
      security:
      - cookieAuth: []
  /api/v1/problem-sets/{shareCode}/expire-invite:
    post:
      tags:
      - Problem Sets
      operationId: expireInvite
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      - name: inviteeName
        in: query
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
      security:
      - cookieAuth: []
...
  /api/v1/problems/{bookSlug}/{code}:
    get:
      tags:
      - Problems
      summary: Get problem by key
      description: Retrieves a specific problem by its book slug and code
      operationId: getProblem
      parameters:
      - name: bookSlug
        in: path
        description: Book slug
        required: true
        schema:
          type: string
      - name: code
        in: path
        description: Problem code
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved problem
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDto"
        "404":
          description: Problem not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDto"
      security: []
...
  /api/v1/problem-sets/{shareCode}:
    get:
      tags:
      - Problem Sets
      operationId: getProblemSetByShareCode
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemSetDto"
      security: []
  /api/v1/problem-sets/{shareCode}/users:
    get:
      tags:
      - Problem Sets
      operationId: getUserRoles
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserNameWithRole"
      security:
      - cookieAuth: []
  /api/v1/problem-sets/{shareCode}/invited-users:
    get:
      tags:
      - Problem Sets
      operationId: getInvitedUsers
      parameters:
      - name: shareCode
        in: path
        required: true
        schema:
          type: string
          minLength: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      security:
      - cookieAuth: []
  /api/v1/problem-sets/public:
    get:
      tags:
      - Problem Sets
      operationId: getPublicProblemSets
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          exclusiveMinimum: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemSetMetadata"
      security: []
  /api/v1/problem-sets/owned:
    get:
      tags:
      - Problem Sets
      operationId: getOwnedProblemSets
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          exclusiveMinimum: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemSetMetadata"
      security:
      - cookieAuth: []
  /api/v1/problem-sets/joined:
    get:
      tags:
      - Problem Sets
      operationId: getJoinedProblemSets
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          exclusiveMinimum: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemSetMetadata"
      security:
      - cookieAuth: []
...
components:
  schemas:
...
    CreateProblemSetRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 0
        description:
          type: string
          maxLength: 255
          minLength: 0
        isPublic:
          type: boolean
        problemKeys:
          type: array
          items:
            type: string
          minItems: 1
      required:
      - description
      - isPublic
      - name
      - problemKeys
    ProblemMetadata:
      type: object
      properties:
        key:
          type: string
        code:
          type: string
        bookSlug:
          type: string
        isHard:
          type: boolean
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
          - SOLVED
          - SOLVING
          - FAILED
          - NOT_SOLVED
      required:
      - bookSlug
      - code
      - isHard
      - key
      - status
      - tags
    ProblemSetDto:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        admins:
          type: array
          items:
            type: string
        isPublic:
          type: boolean
        problems:
          type: array
          items:
            $ref: "#/components/schemas/ProblemMetadata"
        shareCode:
          type: string
      required:
      - admins
      - isPublic
      - name
      - problems
      - shareCode
    ChangeProblemSetProblemsRequest:
      type: object
      properties:
        problemKeys:
          type: array
          items:
            type: string
          minItems: 1
      required:
      - problemKeys
...
    ProblemDto:
      type: object
      properties:
        key:
          type: string
        code:
          type: string
        bookSlug:
          type: string
        isHard:
          type: boolean
        tags:
          type: array
          items:
            type: string
        text:
          type: string
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/Subtask"
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
          - SOLVED
          - SOLVING
          - FAILED
          - NOT_SOLVED
        hasResult:
          type: boolean
      required:
      - bookSlug
      - code
      - hasResult
      - images
      - isHard
      - key
      - status
      - subtasks
      - tags
      - text
    Subtask:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
      required:
      - id
      - text
    UserNameWithRole:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum:
          - USER
          - MODERATOR
          - ADMIN
      required:
      - name
      - role
    ProblemSetMetadata:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        admins:
          type: array
          items:
            type: string
        shareCode:
          type: string
        isPublic:
          type: boolean
        size:
          type: integer
          format: int64
      required:
      - admins
      - isPublic
      - name
      - shareCode
      - size
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT authentication cookie
      name: X-Auth
      in: cookie